+++ tests/test_cct.py
@@ -1,0 +1,93 @@
+"""Test CCT."""
+import math
+from coloraide.everything import ColorAll as Color
+from . import util
+import pytest
+import unittest
+
+
+class TestOhno2013(util.ColorAssertsPyTest):
+    """Test Ohno 2013."""
+
+    @pytest.mark.parametrize(
+        'color',
+        [
+            'green',
+            'yellow',
+            'orange',
+            'white',
+            'gray',
+            'pink',
+
+        ]
+    )
+    def test_cct(self, color):
+        """Test CCT methods."""
+
+        cct, duv = Color(color).cct()
+        c = Color.blackbody(cct, duv)
+        cct2, duv2 = c.cct()
+        assert math.isclose(cct, cct2, abs_tol=0.5)
+        assert math.isclose(duv, duv2, abs_tol=0.5)
+
+
+class TestRobertson1968(util.ColorAssertsPyTest):
+    """Test Robertson 1968."""
+
+    @pytest.mark.parametrize(
+        'color',
+        [
+            'green',
+            'yellow',
+            'orange',
+            'white',
+            'gray',
+            'pink',
+
+        ]
+    )
+    def test_cct(self, color):
+        """Test CCT methods."""
+
+        cct, duv = Color(color).cct(method='robertson-1968')
+        c = Color.blackbody(cct, duv, method='robertson-1968')
+        cct2, duv2 = c.cct(method='robertson-1968')
+        assert math.isclose(cct, cct2, abs_tol=0.5)
+        assert math.isclose(duv, duv2, abs_tol=0.5)
+
+
+class TestCCTSpecificCases(util.ColorAsserts, unittest.TestCase):
+    """Test CCT specific cases."""
+
+    def test_bad_algorithm(self):
+        """Test bad algorithm inputs."""
+
+        with self.assertRaises(ValueError):
+            Color.blackbody(2000, method='bad')
+
+        with self.assertRaises(ValueError):
+            Color('blue').cct(method='bad')
+
+    def test_output_space(self):
+        """Test output space."""
+
+        self.assertColorEqual(
+            Color.blackbody(2000, out_space='display-p3'),
+            Color('color(display-p3 0.93959 0.56696 0.22948)')
+        )
+
+    def test_normalization_space(self):
+        """Test normalization space."""
+
+        self.assertColorEqual(
+            Color.blackbody(2000, space='display-p3'),
+            Color('color(display-p3 1 0.6107 0.25811)')
+        )
+
+    def test_no_normalization_space(self):
+        """Test normalization space."""
+
+        self.assertColorEqual(
+            Color.blackbody(2000, space=None),
+            Color('color(xyz-d65 1.2743 1 0.14524)')
+        )